refactor baseball project
