+++ challenge/220318/src/main/java/baseball/domain/Ball.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 import java.util.Objects;
