+++ challenge/220318/src/main/java/baseball/domain/PlayResult.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 public class PlayResult {
