+++ challenge/220318/src/test/java/baseball/domain/BallTest.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 import org.junit.jupiter.api.Test;
