+++ challenge/220318/src/test/java/baseball/domain/BallsTest.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.Test;
